requirements.txt
alnlp==1.0.0rc27
altair==4.1.0
appnope==0.1.2
argon2-cffi==21.1.0
astor==0.8.1
attrs==21.2.0
Automat==20.2.0
backcall==0.2.0
base58==2.1.1
bcrypt==3.2.0
bleach==4.1.0
blinker==1.4
blis==0.7.5
cachetools==4.2.4
catalogue==2.0.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==7.1.2
constantly==15.1.0
crochet==2.0.0
cryptography==35.0.0
cssselect==1.1.0
cymem==2.0.6
Cython==0.29.24
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
entrypoints==0.3
extra-streamlit-components==0.1.52
filelock==3.3.2
gitdb==4.0.9
GitPython==3.1.24
h2==3.2.0
hanlp==2.1.0a64
hanlp-common==0.0.9
hanlp-downloader==0.0.25
hanlp-trie==0.0.4
hpack==3.0.0
huggingface-hub==0.1.2
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
ipykernel==6.5.0
ipython==7.29.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itemadapter==0.4.0
itemloaders==1.0.4
jedi==0.18.0
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
jsonschema==4.2.1
jupyter-client==7.0.6
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
langcodes==3.3.0
lxml==4.6.4
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
murmurhash==1.0.6
nbclient==0.5.8
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.5
numpy==1.21.4
packaging==21.2
pandas==1.3.4
pandocfilters==1.5.0
parsel==1.6.0
parso==0.8.2
pathy==0.6.1
pexpect==4.8.0
phrasetree==0.0.8
pickleshare==0.7.5
Pillow==8.4.0
preshed==3.0.6
priority==1.3.0
prometheus-client==0.12.0
prompt-toolkit==3.0.22
Protego==0.1.16
protobuf==3.19.1
ptyprocess==0.7.0
py==1.11.0
pyarrow==6.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.8.2
pydeck==0.7.1
PyDispatcher==2.0.5
Pygments==2.10.0
PyJWT==2.3.0
Pympler==0.9
pynvml==11.0.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==22.3.0
queuelib==1.6.2
regex==2021.11.10
requests==2.26.0
retry==0.9.2
sacremoses==0.0.46
Scrapy==2.5.1
Send2Trash==1.8.0
sentencepiece==0.1.96
service-identity==21.1.0
six==1.16.0
smart-open==5.2.1
smmap==5.0.0
spacy-legacy==3.0.8
spacy-loggers==1.0.1
spacy-pkuseg==0.0.28
srsly==2.4.2
streamlit==1.2.0
streamlit-authenticator==0.0.5
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
thinc==8.0.13
tokenizers==0.10.3
toml==0.10.2
toolz==0.11.2
toposort==1.5
torch==1.10.0
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
transformers==4.12.3
Twisted==21.7.0
typer==0.4.0
typing-extensions==3.10.0.2
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
validators==0.18.2
w3lib==1.22.0
wasabi==0.8.2
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
wrapt==1.13.3
zope.interface==5.4.0
